Support recent versions of bytestring using safer abstraction.
The old code was for an earlier bytestring version which would
retain the same foreign pointer and use offsets for slicing.

This wasn't safe, and also isn't true anymore.

Move to a boxed vector, which I think might actually be faster,
as the sort only needs to swap 1 pointer instead of 6 unboxed
values.